backend/requirements.txt
aniso8601==9.0.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-RESTful==0.3.9
graphqlclient==0.2.4
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
oauthlib==3.2.1
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
